Arreglos Warnings Peleas y BarraVida
Arreglo de los warnings en ambos scripts

+++ DeathTokenBase/DeathToken/BarraVida.h
@@ -21,7 +21,7 @@ public:
 	void establecerValor(int nuevoValor);
 	void establecerPosicion(int x, int y);
 	void establecerVisible(bool esVisible);
-	void updateColorBasedOnHealth(float currentHealth, float maxHealth);
+	void updateColorBasedOnHealth(float currentHealth, int maxHealth);
 	int obtenerValor() const;
 	bool estaVisible() const;
